[Refactor] create components for CounselorCenter PersonalInfo
