+++ package.json
@@ -16,6 +16,7 @@
     "@types/react-slick": "^0.23.10",
     "@types/swiper": "^6.0.0",
     "antd": "^5.3.1",
+    "antd-img-crop": "^4.12.2",
     "axios": "^1.3.5",
     "cookies-next": "^2.1.1",
     "dayjs": "^1.11.7",
